Refactor, clean up
