+++ build-system/compile/compile.js
@@ -16,29 +16,23 @@
 'use strict';
 
 const argv = require('minimist')(process.argv.slice(2));
-const closureCompiler = require('google-closure-compiler');
 const colors = require('ansi-colors');
 const fs = require('fs-extra');
 const gulp = require('gulp');
 const nop = require('gulp-nop');
 const rename = require('gulp-rename');
 const rimraf = require('rimraf');
-const shortenLicense = require('../shorten-license');
+const shortenLicense = require('./shorten-license');
 const sourcemaps = require('gulp-sourcemaps');
-const {highlight} = require('cli-highlight');
+const {gulpClosureCompile, handleCompilerError} = require('./closure-compile');
 const {isTravisBuild} = require('../travis');
-const {singlePassCompile} = require('../single-pass');
+const {singlePassCompile} = require('./single-pass');
 const {VERSION: internalRuntimeVersion} = require('../internal-version') ;
 
 const isProdBuild = !!argv.type;
 const queue = [];
 let inProgress = 0;
 const MAX_PARALLEL_CLOSURE_INVOCATIONS = 4;
-
-// Also used in gulpfile.js and single-pass.js
-const CHECK_TYPES_NAILGUN_PORT = '2114';
-const DIST_NAILGUN_PORT = '2115';
-
 
 // Compiles AMP with the closure compiler. This is intended only for
 // production use. During development we intend to continue using
@@ -86,18 +80,6 @@ function cleanupBuildDir() {
   fs.mkdirsSync('build/fake-polyfills/src/polyfills');
 }
 exports.cleanupBuildDir = cleanupBuildDir;
-
-// Formats a closure compiler error message into a more readable form by
-// dropping the closure compiler plugin's logging prefix and then syntax
-// highlighting the error text.
-function formatClosureCompilerError(message) {
-  const closurePluginLoggingPrefix = /^.*?gulp-google-closure-compiler.*?: /;
-  message = message.replace(closurePluginLoggingPrefix, '');
-  message = highlight(message, {ignoreIllegals: true});
-  message = message.replace(/WARNING/g, colors.yellow('WARNING'));
-  message = message.replace(/ERROR/g, colors.red('ERROR'));
-  return message;
-}
 
 function compile(entryModuleFilenames, outputDir, outputFilename, options) {
   const hideWarningsFor = [
@@ -347,7 +329,7 @@ function compile(entryModuleFilenames, outputDir, outputFilename, options) {
     externs.push('build-system/amp.multipass.extern.js');
 
     /* eslint "google-camelcase/google-camelcase": 0*/
-    let compilerOptions = {
+    const compilerOptions = {
       compilation_level: options.compilationLevel || 'SIMPLE_OPTIMIZATIONS',
       // Turns on more optimizations.
       assume_function_wrapper: true,
@@ -406,66 +388,32 @@ function compile(entryModuleFilenames, outputDir, outputFilename, options) {
       delete compilerOptions.define;
     }
 
-    let compilerErrors = '';
-    const initOptions = {
-      extraArguments: ['-XX:+TieredCompilation'], // Significant speed up!
-    };
-    const pluginOptions = {
-      platform: ['java'], // Override the binary used by closure compiler
-      logger: errors => compilerErrors = errors, // Capture compiler errors
-    };
-
-    // SIGNIFICANTLY speed up compilation on Mac and Linux using nailgun
-    // See https://github.com/facebook/nailgun.
-    if (process.platform == 'darwin' || process.platform == 'linux') {
-      const compilerOptionsArray = [
-        '--nailgun-port',
-        options.typeCheckOnly ? CHECK_TYPES_NAILGUN_PORT : DIST_NAILGUN_PORT,
-        'org.ampproject.AmpCommandLineRunner',
-        '--',
-      ];
-      Object.keys(compilerOptions).forEach(function(option) {
-        const value = compilerOptions[option];
-        if (value instanceof Array) {
-          value.forEach(function(item) {
-            compilerOptionsArray.push('--' + option + '=' + item);
-          });
+    const compilerOptionsArray = [];
+    Object.keys(compilerOptions).forEach(function(option) {
+      const value = compilerOptions[option];
+      if (value instanceof Array) {
+        value.forEach(function(item) {
+          compilerOptionsArray.push('--' + option + '=' + item);
+        });
+      } else {
+        if (value != null) {
+          compilerOptionsArray.push('--' + option + '=' + value);
         } else {
-          if (value != null) {
-            compilerOptionsArray.push('--' + option + '=' + value);
-          } else {
-            compilerOptionsArray.push('--' + option);
-          }
+          compilerOptionsArray.push('--' + option);
         }
-      });
-      compilerOptions = compilerOptionsArray; // nailgun-runner takes an array
-      pluginOptions.platform = ['native']; // nailgun-runner isn't a java binary
-      initOptions.extraArguments = null; // Already part of nailgun-server
-    }
+      }
+    });
 
-    // Override to local closure compiler JAR
-    closureCompiler.compiler.JAR_PATH =
-        require.resolve('../runner/dist/runner.jar');
-
-    const handleCompilerError = function(err) {
-      console./*OK*/error(colors.red(
-          'Compilation failed for ' + outputFilename + ':\n') +
-          formatClosureCompilerError(compilerErrors) + '\n' +
-          err);
-      process.exit(1);
-    };
-
-    const compiler = closureCompiler.gulp(initOptions);
     if (options.typeCheckOnly) {
       return gulp.src(srcs, {base: '.'})
-          .pipe(compiler(compilerOptions, pluginOptions))
+          .pipe(gulpClosureCompile(compilerOptionsArray))
           .on('error', handleCompilerError)
           .pipe(nop())
           .on('end', resolve);
     } else {
       return gulp.src(srcs, {base: '.'})
           .pipe(sourcemaps.init({loadMaps: true}))
-          .pipe(compiler(compilerOptions, pluginOptions))
+          .pipe(gulpClosureCompile(compilerOptionsArray))
           .on('error', handleCompilerError)
           .pipe(rename(outputFilename))
           .pipe(sourcemaps.write('.'))
