+++ build-system/compile/shorten-license.js
